Refactor: ajout de fichiers js pour séparer les fonctionnalités
