+++ FrontEnd/assets/pages/login.html
@@ -35,7 +35,7 @@
       <section id="login">
         <h2>Log in</h2>
 
-        <form action="#" method="post">
+        <form method="post">
           <label for="email">Email</label>
           <input type="email" name="email" id="email" />
