+++ FrontEnd/index.html
@@ -16,6 +16,7 @@
     ></script>
     <script type="module" src="js/filtre.js" defer></script>
     <script type="module" src="js/edit.js" defer></script>
+    <script type="module" src="js/main.js" defer></script>
     <meta name="description" content="" />
     <link rel="stylesheet" href="./assets/css/style.css" />
     <link rel="stylesheet" href="./assets/css/filtre.css" />
